IT FUCKING WORKS!!!!!
pardon my french, however, I have spent far too long on this to not be excited. It works to the extent that the cost goes down aproaching 0 and the proper position in the nn increases while the rest decrease. this is good. in this case, I'm only training it off of the 1st training example in the set, which is a 7, so it is very good at saying everything is a 7. I'm going to now work on some optimization and getting it to learn all the numbers.