maybe
I'm getting the matrix stuff to work, but I', not sure how to deal with relu or lrelu cuz I don't know what I'd set the ideal value for the output to. on the bright side I was able to get a foreprop function in 3 lines of code! I'm going to try not doing any hidden layers right now. I can still get 90-93% people have been saying so we'll see.